+++ src/properties/OrbitalEnergies.h
@@ -67,7 +67,7 @@ public:
         println(0, o_head.str());
         mrcpp::print::separator(0, '-');
 
-        for (size_t i = 0; i < this->epsilon.size(); i++) {
+        for (int i = 0; i < this->epsilon.size(); i++) {
             auto sp = 'u';
             if (this->spinVec(i) == SPIN::Paired) sp = 'p';
             if (this->spinVec(i) == SPIN::Alpha) sp = 'a';
@@ -88,7 +88,7 @@ public:
         DoubleVector occ = getOccupation();
         const DoubleVector &eps = getEpsilon();
         std::vector<std::string> spn;
-        for (size_t i = 0; i < spinVec.size(); i++) {
+        for (int i = 0; i < spinVec.size(); i++) {
             if (this->spinVec(i) == SPIN::Paired) spn.push_back("p");
             if (this->spinVec(i) == SPIN::Alpha) spn.push_back("a");
             if (this->spinVec(i) == SPIN::Beta) spn.push_back("b");
