+++ src/qmfunctions/orbital_utils.cpp
@@ -418,7 +418,7 @@ void orbital::orthogonalize(double prec, Orbital &&phi, Orbital psi) {
 void orbital::orthogonalize(double prec, OrbitalVector &Phi) {
     mrcpp::mpi::free_foreign(Phi);
     for (size_t i = 0; i < Phi.size(); i++) {
-        for (int j = 0; j < i; j++) {
+        for (size_t j = 0; j < i; j++) {
             int tag = 7632 * i + j;
             int src = (Phi[j].getRank()) % mrcpp::mpi::wrk_size;
             int dst = (Phi[i].getRank()) % mrcpp::mpi::wrk_size;
@@ -747,7 +747,7 @@ IntVector orbital::get_spins(const OrbitalVector &Phi) {
  *
  */
 void orbital::set_spins(OrbitalVector &Phi, const IntVector &spins) {
-    if (Phi.size() != spins.size()) MSG_ERROR("Size mismatch");
+    if (Phi.size() != static_cast<size_t>(spins.size())) MSG_ERROR("Size mismatch");
     for (size_t i = 0; i < Phi.size(); i++) Phi[i].spin() = i;
 }
 
@@ -765,7 +765,7 @@ DoubleVector orbital::get_occupations(const OrbitalVector &Phi) {
  *
  */
 void orbital::set_occupations(OrbitalVector &Phi, const DoubleVector &occup) {
-    if (Phi.size() != occup.size()) MSG_ERROR("Size mismatch");
+    if (Phi.size() != static_cast<size_t>(occup.size())) MSG_ERROR("Size mismatch");
     for (size_t i = 0; i < Phi.size(); i++) Phi[i].occ() = occup(i);
 }
 
@@ -879,7 +879,7 @@ void orbital::print(const OrbitalVector &Phi) {
 }
 
 DoubleVector orbital::calc_eigenvalues(const OrbitalVector &Phi, const ComplexMatrix &F_mat) {
-    if (F_mat.cols() != Phi.size()) MSG_ABORT("Invalid Fock matrix");
+    if (static_cast<size_t>(F_mat.cols()) != Phi.size()) MSG_ABORT("Invalid Fock matrix");
     if (not orbital::orbital_vector_is_sane(Phi)) MSG_ABORT("Insane orbital vector");
 
     DoubleVector epsilon = DoubleVector::Zero(Phi.size());
