+++ src/qmoperators/two_electron/XCPotential.cpp
@@ -72,7 +72,7 @@ void XCPotential::setup(double prec) {
         std::vector<double> prefacts;
         std::vector<double> nlccrs;
         // Eigen::VectorXd nlccqcores;
-        for (int i = 0; i < this->nucs->size(); i++) {
+        for (size_t i = 0; i < this->nucs->size(); i++) {
             Nucleus nuc = (*nucs)[i];
             if ( nuc.hasPseudopotential()) {
                 if (nuc.getPseudopotentialData()->getHasNlcc()) {
@@ -88,9 +88,9 @@ void XCPotential::setup(double prec) {
         bool hasnlcc = prefacts.size() > 0;
         if (hasnlcc) {
             auto rho_analytic = [prefacts, nlccCoords, nlccrs](const mrcpp::Coord<3> &r) {
-                int n = prefacts.size();
+                size_t n = prefacts.size();
                 double rho = 0.0;
-                for (int i = 0; i < n; i++) {
+                for (size_t i = 0; i < n; i++) {
                     double rr = std::sqrt((r[0] - nlccCoords[i][0]) * (r[0] - nlccCoords[i][0])
                         + (r[1] - nlccCoords[i][1]) * (r[1] - nlccCoords[i][1])
                         + (r[2] - nlccCoords[i][2]) * (r[2] - nlccCoords[i][2]));
@@ -185,7 +185,7 @@ Density &XCPotential::getDensity(DensityType spin, int pert_idx) {
         NOT_IMPLEMENTED_ABORT;
     }
     if (dens_idx < 0) MSG_ABORT("Invalid density index");
-    if (dens_idx > densities.size()) MSG_ABORT("Invalid density index");
+    if (dens_idx > static_cast<int>(densities.size())) MSG_ABORT("Invalid density index");
     return densities[dens_idx];
 }
